:recycle: Separate ViewGraphSetVisitor
